fixes small issues
