+++ lucida/musicservice/README.md~
@@ -6,8 +6,7 @@ This is a mood-indexed music scanning service.
 - [Gracenote Music API](https://www.gracenote.com/)
 
 ## Build
-- get API key for [Gracenote Music API](https://developer.gracenote.com/user/me/apps)
-- place client ID and key in `MusicConfig.py`
+- Get API key for [Gracenote Music API](https://developer.gracenote.com/user/me/apps)
 ```
 make all
 ```
